feat(endpoints/bid/model): add Tenders table
rename Bid table -> MatchResult
